Allow setting the ruleset from the command line
